refactor(handler) extract parse_http_req_headers to module
