+++ kong/plugins/zipkin/handler.lua
@@ -1,10 +1,10 @@
 local new_zipkin_reporter = require "kong.plugins.zipkin.reporter".new
 local new_span = require "kong.plugins.zipkin.span".new
 local to_hex = require "resty.string".to_hex
+local parse_http_req_headers = require "kong.plugins.zipkin.parse_http_req_headers"
 
 local subsystem = ngx.config.subsystem
 local fmt = string.format
-local char = string.char
 
 local ZipkinLogHandler = {
   VERSION = "0.2.1",
@@ -16,89 +16,6 @@ local ZipkinLogHandler = {
 local reporter_cache = setmetatable({}, { __mode = "k" })
 
 local math_random = math.random
-
-local baggage_mt = {
-  __newindex = function()
-    error("attempt to set immutable baggage", 2)
-  end,
-}
-
-
-local function hex_to_char(c)
-  return char(tonumber(c, 16))
-end
-
-
-local function from_hex(str)
-  if str ~= nil then -- allow nil to pass through
-    str = str:gsub("%x%x", hex_to_char)
-  end
-  return str
-end
-
-
-local function parse_http_headers(headers)
-  local warn = kong.log.warn
-  -- X-B3-Sampled: if an upstream decided to sample this request, we do too.
-  local should_sample = headers["x-b3-sampled"]
-  if should_sample == "1" or should_sample == "true" then
-    should_sample = true
-  elseif should_sample == "0" or should_sample == "false" then
-    should_sample = false
-  elseif should_sample ~= nil then
-    warn("x-b3-sampled header invalid; ignoring.")
-    should_sample = nil
-  end
-
-  -- X-B3-Flags: if it equals '1' then it overrides sampling policy
-  -- We still want to warn on invalid sample header, so do this after the above
-  local debug = headers["x-b3-flags"]
-  if debug == "1" then
-    should_sample = true
-  elseif debug ~= nil then
-    warn("x-b3-flags header invalid; ignoring.")
-  end
-
-  local had_invalid_id = false
-
-  local trace_id = headers["x-b3-traceid"]
-  if trace_id and ((#trace_id ~= 16 and #trace_id ~= 32) or trace_id:match("%X")) then
-    warn("x-b3-traceid header invalid; ignoring.")
-    had_invalid_id = true
-  end
-
-  local parent_id = headers["x-b3-parentspanid"]
-  if parent_id and (#parent_id ~= 16 or parent_id:match("%X")) then
-    warn("x-b3-parentspanid header invalid; ignoring.")
-    had_invalid_id = true
-  end
-
-  local span_id = headers["x-b3-spanid"]
-  if span_id and (#span_id ~= 16 or span_id:match("%X")) then
-    warn("x-b3-spanid header invalid; ignoring.")
-    had_invalid_id = true
-  end
-
-  if trace_id == nil or had_invalid_id then
-    return nil
-  end
-
-  local baggage = {}
-  trace_id = from_hex(trace_id)
-  parent_id = from_hex(parent_id)
-  span_id = from_hex(span_id)
-
-  -- Process jaegar baggage header
-  for k, v in pairs(headers) do
-    local baggage_key = k:match("^uberctx%-(.*)$")
-    if baggage_key then
-      baggage[baggage_key] = ngx.unescape_uri(v)
-    end
-  end
-  setmetatable(baggage, baggage_mt)
-
-  return trace_id, span_id, parent_id, should_sample, baggage
-end
 
 
 local function get_reporter(conf)
@@ -171,7 +88,7 @@ if subsystem == "http" then
     local req = kong.request
 
     local trace_id, span_id, parent_id, should_sample, baggage =
-      parse_http_headers(req.get_headers())
+      parse_http_req_headers(req.get_headers())
     local method = req.get_method()
 
     if should_sample == nil then
